overall refactor to match figma design of news section
