+++ src/components/Slider/styles/Slider.module.scss
@@ -1,0 +1,22 @@
+figure {
+  max-height: 500px;
+  margin: 0;
+}
+
+button {
+  margin: auto 0;
+  background: none !important;
+  opacity: 1 !important;
+}
+
+.content {
+  h1 {
+    font-weight: 700;
+    font-size: 40px;
+  }
+  font-family: Mulish;
+  font-weight: 400;
+  font-size: 20px;
+  max-width: 684px;
+  margin: auto;
+}
