+++ src/pages/News/styles/News.component.scss
@@ -1,0 +1,5 @@
+h1 {
+  text-align: center;
+  font-weight: 600;
+  font-size: 48px;
+}
